+++ lib/linkage/Utils.hh
@@ -51,7 +51,7 @@ Glib::ustring suffix_value(float value);
 Glib::ustring str(int value);
 Glib::ustring str(unsigned int value);
 Glib::ustring str(size_type value);
-Glib::ustring str(size_t value);
+Glib::ustring str(unsigned long value);
 Glib::ustring str(float value, int precision);
 Glib::ustring str(const sha1_hash& hash);
 Glib::ustring get_eta(size_type size, float rate);
